requirements.txt
asttokens==2.4.1
decorator==5.1.1
et-xmlfile==1.1.0
executing==2.1.0
generate-dummyf==1.1.4
ipdb==0.13.13
ipython==8.27.0
jedi==0.19.1
lxml==5.3.0
matplotlib-inline==0.1.7
numpy==2.1.1
openpyxl==3.1.5
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
SQLAlchemy==2.0.34
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
wcwidth==0.2.13
